Put NFoldMetric in product_riemannian_metric
